Get rid of RunManager
